Farewell Play, this is going to be a standalone JAR
Given the memory constraints of the conversion, a dedicated server to
handle the conversion automatically would be quite expensive. It's not
worth it, since the need for this is so infrequent. Instead, we'll have
the files get converted before being uploaded to the server

Also updated tests that were left behind when we started excluding
polygons without a material